Get CrudApplication out of the Activity's "application" instance.
The purpose is to eliminate the need for defining a CrudType and Activity classes for each EntityType.
This would greatly simplify application code and how much needs to be understood to control behavior.

Added CrudAndroidApplication to extend anroid.app.Application and wrap the CrudApplication.
Moved defaultContentUri to CrudApplication and use the first EntityType as the primary one.
Added UriPath.lastEntityNameOption.