Move apklib contents to scrud-android-res to be built by Maven.
This is to work around sbt not supporting apklibs.
The plan is to use Maven as needed (rarely) to build the apklib contents.
Then use sbt to build a simple jar for scrud-android which depends on the scrud-android-res.apklib.
Then downstreams can use both to build apps (which are supported by sbt).